# contracts/openapi/bootstrap.yaml
openapi: 3.1.0
...
info:
  title: O3K Bootstrap API
  version: 0.2.0-alpha.1
  description: >-
    Bootstrap-only contract. It does not claim complete OpenStack compatibility.
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /v3/auth/projects:
    get:
      operationId: listAuthProjects
      summary: List the projects the authenticated subject may scope into
      description: >-
        Client project discovery used by Horizon's project switcher. Accepts an
        unscoped or project-scoped token and returns only the caller's durable
        role-assignment projects. It is not project administration.
      responses:
        '200':
          description: Caller's role-assignment projects.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorizedProjects'
        '401':
          description: Missing, invalid, or expired token.
        '503':
          description: Bootstrap identity is not configured.
...
components:
  schemas:
...
    AuthorizedProjects:
      type: object
      required: [projects]
      properties:
        projects:
          type: array
          items:
            type: object
            required: [id, name, domain_id, enabled]
            properties:
              id: {type: string}
              name: {type: string}
              domain_id: {type: string}
              enabled: {type: boolean}
...